updates on feedback
