+++ src/pyramid_debugtoolbar/toolbar.py
@@ -62,8 +62,8 @@ class DebugToolbar(object):
         self.default_active_panels = default_active_panels
         self.visible = False
 
-        # Panels can be be activated (more features) (e.g. Performace panel)
-        # toolbar.js controls this cookie with the following concepts
+        # Panels can be be activated (more features) (e.g. Performance panel)
+        # toolbar.js controls this cookie with the following concepts:
         #   COOKIE_NAME_ACTIVE = 'pdtb_active'
         #   COOKIE_DELIM_ACTIVE = ','
         pdtb_active = url_unquote(request.cookies.get('pdtb_active', ''))
